ImagetoSpeech/requirements.txt
google-generativeai
gTTS
langdetect
deep-translator